Add blacklist to notifier
We want to ensure that some users never get emailed about whether their
data sets are out of date.

This sets up a file called blacklist.json which will be used inside of
the emailer instance to block emails being sent to anyone in that list.

This ignores case difference between the blacklist and the email address
that is being sent to.