Update template to incorporate dashboards
We want to list the dashboards that a particular data set is used on to
give the end user a little more context that a name that they have never
seen.

In this commit I have switched the Message 'class' over to using
Mustache as the string concat was crazy.